Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21124 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/particles/object/particles.cpp
@@ -248,7 +248,7 @@ csParticlesObject::csParticlesObject (csParticlesFactory* p)
 
   scfiRenderBufferAccessor.AttachNew (new eiRenderBufferAccessor (this));
 
-  bufferHolder->SetAccessor (scfiRenderBufferAccessor, ~0);  
+  bufferHolder->SetAccessor (scfiRenderBufferAccessor, (uint32)~0);
 
   csShaderVariable* sv;
   sv = svcontext->GetVariableAdd (radius_name);
